libs/platypus-core/src/lib/domain/common/validators/required-field.validator.ts
import { DataUtils } from '../../../boundaries/utils';
import {
  ValidationRule,
  ValidatorResult,
} from '../../../boundaries/validation';

export class RequiredFieldValidator extends ValidationRule {
  validate(field: string, value: unknown): ValidatorResult {
    if (
      !value ||
      (Array.isArray(value) && !value.length) ||
      (DataUtils.isString(value) && !(value as string).trim())
    ) {
      return {
        valid: false,
        errors: {
          name: field + ' is required field',
        },
      } as ValidatorResult;
    }

    return { valid: true, errors: {} };
  }
}
